+++ api/src/main/java/org/allaymc/api/entity/component/EntityCommandSenderComponent.java
@@ -1,7 +1,0 @@
-package org.allaymc.api.entity.component;
-
-/**
- * @author daoge_cmd
- */
-public interface EntityCommandSenderComponent {
-}