+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityPainting.java
@@ -1,7 +1,7 @@
 package org.allaymc.api.entity.interfaces;
 
-import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityPaintingBaseComponent;
 
-public interface EntityPainting extends Entity {
+public interface EntityPainting extends EntityLiving, EntityPaintingBaseComponent {
 
 }